Setup plugin loading on remote.
